refactor & fix double write
