+++ internal/backend/github_actions_cache.go
@@ -247,13 +247,6 @@ func (c *GitHubActionsCache) Get(ctx context.Context, objectID string, w io.Writ
 }
 
 func (c *GitHubActionsCache) Put(ctx context.Context, objectID string, size int64, r io.Reader) error {
-	defer func() {
-		_, err := io.Copy(io.Discard, r)
-		if err != nil {
-			c.logger.Warnf("discard body: %v", err)
-		}
-	}()
-
 	key, _ := c.objectBlobKey(objectID)
 	return c.storeCache(ctx, key, size, r)
 }
